requirements.txt
openpyxl
requests
Pillow
beautifulsoup4
rich
urllib3